Make writing out work again, using the information from the PVD.
First we had to fix a bug, where the path table records
were being written to the wrong place.  Then we had to
remove all of the tracking of location in write(), since
we now do that as we are adding things.  The result is
shorter and easier to understand code in write(), and
more correct.

Signed-off-by: Chris Lalancette <clalancette@irobot.com>