Get rid of find_parent_dirnum.
This was a totally unnecessary (and unnecessarily slow)
function.  First of all, it was really a function that
belonged as part of the DirectoryRecord, not the headervd.
Second, we really didn't need it at all since we now
have the PTR attached to the directory record; we can
just reach directly there and get the number we need.
Remove it with prejudice.

Signed-off-by: Chris Lalancette <clalancette@irobot.com>