Optimize adding children to a directory.
On ISOs with a large number of children in a single
directory, we were spending way too much time adding
a child.  The problem was that we were doing a linear
search through the children, checking for duplicate
children file names.  As the number of children grew
large, that linear search grew expensive.  Instead,
we recognize that bisect.bisect_left will always
choose an index to the left of a duplicate child.
Thus, to check for duplicates, we only need to see
if we are duplicate with the child that is currently
at the index that bisect returns.  This significantly
speeds up adding children to a directory.  We also add a
test for this case.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>