Fix a bug where we were mis-assigning ptr records.
If two subdirectories at any level had the same name,
we were only assigning a PTR record to the first of
them, because our dictionary used just the name as the
key.  Instead, use the name+directory number as the key,
which uniquely identifies everything on the ISO.

Also add in unit tests for this.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>