Don't use string concatenation.
In Python, strings are immutable objects.  This mean
that when concatenating them (via +), new objects have
to be created every time.  Even worse, if there are
intermediate strings, new objects are created for
each of those.  Avoid this whole thing by using string
formatting where possible, and string join on lists
where we don't know the size of the list a priori.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>