Make it so that we can better deal with hidden entries.
Particularly around El Torito entries.  To do this, we
add in the concept of a "hidden" directory record.  These
are records that are used mostly for containing the file
objects appropriately.  Now that we have that concept,
we put hidden directory records in when a file would
disappear from the filesystem completely.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>