Specifically exclude symlinks from linked_records.
It is not enough for us to just exclude zero-length
files, since FreeBSD, at least, gives a data-length
to symlinked files.  The Rock Ridge specification
is not clear on whether symlinked files should have
a data length or not.  genisoimage assigns them a
zero size; FreeBSD seems to assign them the size
of the symlink.  In any case, we allow for the
FreeBSD style here.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>