Rewrite printing mode support in pycdlib-explorer.
Instead of having two separate toggles, which toggle each
other (and didn't have a way to print out the current
mode), instead have a single toggle that controls whether
we are printing in iso9660, rock ridge, or joliet mode.
While we are at it, fix a couple of bugs found in the
mode support.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>