Set the app_ident_str default a different way.
Previously, we were setting it as the "default" for
the named argument.  The problem with that is that if
the user passes in a blank string (which is pretty
common for "use the default"), then the app id would
have been blank, not the default.  Doing it this
way ensures we get the default on a blank string,
which is slightly better for users of the API.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>