Heavily refactor Rock Ridge Continuation Entry.
The previous implementation was a bit ad-hoc.  In
particular, Rock Ridge Continuation Entries are organized
into shared blocks, but this wasn't really reflected
in the previous implementation.  Now, we track Rock
Ridge Continuation "Blocks" at the PVD level; this
allows them to be conceptually shared amongst Rock
Ridge Continuation Entries.  With that in place, we
can now track Rock Ridge entries in a more abstract
way and only assign them to a block when we go
through reassign_extents.  Further, we know exactly
when we allocated a new block so we can add size to
the PVD as necessary.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>