Optimize update_ptr_records.
We recognize that updating the directory record extents
can be done inline in reassign_vd_directory_records(),
which means we can pull that particular line out of
update_ptr_records().  Once that is gone, we recognize
that we only ever need to update the directory numbers
when we add or remove a PTR entry, and that only needs
to be done at add_directory/rm_directory time, not
reassign_vd_directory_records().  By doing this, we
cut the amount of time we spend in update_ptr_records()
in half.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>