Simplify the headervd class.
In particular, remove the ident_to_ptr dictionary.  While
we originally did it to speed things up, it turns
out that we can get almost the same speed by actually
utilizing the fact that we are using a sorted list and
using bisect on that list.  By doing this, we remove
a bunch of code, and no longer have a duplicated structure
for path table records.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>