Deal with ISOs with weird XA settings.
The Windows 98 SE ISO is a bit strange.  For each
directory record, there is an XA record at the end
of the size, with a bunch of padding in between.  I
don't know any reason for this in the standards, but
in order to comply with this, we end up adding additional
code that can deal with these weird ISOs.  Note that we
only preserve this odd behavior for existing entries;
any new ones that the user adds will be more standard.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>