Optimize away dot and dotdot directories.
Since we know the directory record list is in sorted
order, we know that dot and dotdot are the first two
entries.  When searching the list, we can just ignore
these, and it turns out that it is generally faster
to slice them away at the beginning than to check for
them in every iteration of the loop.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>