Fix a nasty bug with removal of RR deep relocation.
The bug came about because there was still a part of
the code where we relying on extents to find a child
link record from the target.  Because of this, it wouldn't
work properly in our new lazy extent allocation world.
To fix this, at parse and new time we now store the
link from the child_link to the moved record, and from
the moved record back to the child_link.  This lets us
find it much faster, and fixes the bug.  This also lets
us get rid of a large function, which is great.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>