Fix ISOs that have incorrect PVD sizes.
We have seen ISOs in the wild (Office XP) that have an
incorrect size for the PVD.  That is, there is data linked
to by directory records that is beyond the size specified
in the PVD space_size field.  In these cases, we adjust
the PVD (and SVD, if necessary) space_size field during
open to cover the actual size of the ISO; if we go to
write it out, it will then have the correct size.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>