Start adding UDF support to modify_in_place.
While we are here, there are actually two bugs and a
performance problem here.  For the first bug, we were
re-using the iso object in the test_hybrid modify_file_in_place
tests, which meant we weren't seeing the ISO as it was
on disk, but rather the metadata updates in memory.  This
was masking the second bug, which was that we were
failing to update the Joliet records when modifying
in place.  For the inefficieny we were iterating over
the entire parent of the child looking for the child
and computing the extent and offset it was on.  However,
we are already doing this work inside of DirectoryRecord,
so reuse that information.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>