Increase read size to 2048.
We've seen UDF ISOs in the wild (GT4 PS2 DVD) that
have 2048 bytes for the UDF anchors and file set data.
Since it doesn't cost us much to do it (other than
slightly more time to compute the CRC with lots of
zeros), just unconditionally up the size we read.
While we are in here, do a micro-optimization to
read two consecutive extents together for a small
speedup.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>