Switch to using an internal self.logical_block_size everywhere.
The old method of using self.pvd.logical_block_size() is less
performant and led to us sprinkiling lbs = self.pvd.logical_block_size()
throughout the code.  We also don't have any tests (or
any reports from the wild) of the PVD size disagreeing with
the SVD size.  We also don't have any tests for that, so it
is almost certainly broken.   Finally, with upcoming UDF support it is
going to be possible to have and to test non-2048 block size
support, so it is just easier to have one place to set that.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>