Fix Win32 Raw Devices with a size greater than 4GB~
The struct unpacked only 4 bytes (unsigned long) worth of data for the Disk Size, and left the "Extra Data" as-is, unused. It seems this "Extra Data" is essentially an extension to the data size for 64-bit systems.

You can just use the 4 bytes only, if it's for a device or ISO roughly 4GB~ or less, but if it's more, it would need to unpack all 8 bytes after the DISK_GEOMETRY.

It is also important for it to be `<` and not the default `@` (which nowadays may be `>`).

Without this change, any device that's 5gb~ or more (give or take) will not have the correct total disk size, therefore all SEEK_END or len() calls will fail.